Add API id and structure
